fix(tmserver): drop the redundant uint16 conversions (unconvert)
motionRefineFailBare is already declared uint16, so converting it again is a
no-op that golangci-lint's unconvert rejects. It failed the Lint job on main
while test and docker passed — the linter was the one gate I could not run
locally, which is exactly why it slipped through.

golangci-lint is installed here now. Note that running it on Windows reports
spurious gofmt failures for every file the checkout wrote with CRLF
(core.autocrlf=true); the repository itself stores LF, so CI does not see them.

+++ tmserver/internal/handler/refine_test.go
@@ -547,7 +547,7 @@ func TestRefineFailEmotionDependsOnFace(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got := uint16(motionRefineFailBare)
+			got := motionRefineFailBare
 			if tt.equip/10 != 0 {
 				got = motionRefineFailFaced
 			}
